+++ src/nvim/change.c
@@ -42,7 +42,6 @@
 #include "nvim/os/time.h"
 #include "nvim/plines.h"
 #include "nvim/pos.h"
-#include "nvim/screen.h"
 #include "nvim/search.h"
 #include "nvim/state.h"
 #include "nvim/strings.h"
