+++ src/nvim/drawline.h
@@ -8,7 +8,6 @@
 #include "nvim/decoration_provider.h"
 #include "nvim/fold.h"
 #include "nvim/macros.h"
-#include "nvim/screen.h"
 #include "nvim/types.h"
 
 // Maximum columns for terminal highlight attributes
