+++ src/nvim/drawscreen.h
@@ -22,6 +22,11 @@ enum {
 /// ('lines' and 'rows') must not be changed.
 EXTERN bool updating_screen INIT(= 0);
 
+EXTERN match_T screen_search_hl INIT(= { 0 });       // used for 'hlsearch' highlight matching
+
+#define W_ENDCOL(wp)   ((wp)->w_wincol + (wp)->w_width)
+#define W_ENDROW(wp)   ((wp)->w_winrow + (wp)->w_height)
+
 #ifdef INCLUDE_GENERATED_DECLARATIONS
 # include "drawscreen.h.generated.h"
 #endif
