+++ src/nvim/getchar.c
@@ -54,7 +54,6 @@
 #include "nvim/os/os.h"
 #include "nvim/plines.h"
 #include "nvim/pos.h"
-#include "nvim/screen.h"
 #include "nvim/state.h"
 #include "nvim/strings.h"
 #include "nvim/types.h"
