+++ src/nvim/insexpand.c
@@ -54,7 +54,6 @@
 #include "nvim/popupmenu.h"
 #include "nvim/pos.h"
 #include "nvim/regexp.h"
-#include "nvim/screen.h"
 #include "nvim/search.h"
 #include "nvim/spell.h"
 #include "nvim/state.h"
